test_task_inn/backend/requirements.txt
amqp==5.0.6
appdirs==1.4.4
asgiref==3.4.1
attrs==21.2.0
billiard==3.6.4.0
celery==5.2.1
certifi==2021.10.8
charset-normalizer==2.0.8
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
Deprecated==1.2.13
distlib==0.3.1
Django==3.1.7
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-filter==21.1
django-redis==5.0.0
django-rest-swagger==2.2.0
django-timezone-field==4.2.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
drf-spectacular==0.13.2
filelock==3.0.12
idna==3.3
importlib-metadata==4.8.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
jsonschema==3.2.0
kombu==5.2.2
Markdown==3.3.6
MarkupSafe==2.0.1
openapi-codec==1.3.2
Pillow==8.1.2
postgres==4.0
prompt-toolkit==3.0.23
psycopg2-binary==2.8.6
psycopg2-pool==1.1
PyJWT==2.3.0
pyrsistent==0.18.0
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
redis==4.0.2
requests==2.26.0
simplejson==3.17.6
six==1.15.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.7
vine==5.0.0
virtualenv==20.4.2
wcwidth==0.2.5
wrapt==1.13.3
zipp==3.6.0
